p4i integration Integrate change 65205 from 3.0 to main. Description for change 65205:
: Change 65205 by vohi@vohi-hokkaido-nt on 2002/06/05 10:46:55
: 
: 	Compile.
: 
: Affected files ...
: 
: ... //depot/qt/3.0/src/tools/qcomlibrary.cpp#14 edit

[git-p4: depot-paths = "//depot/qt/main/": change = 65206]

+++ src/tools/qcomlibrary.cpp
@@ -167,8 +167,8 @@ static int qt_tokenize( const char *s, ulong s_len, ulong *advance,
 
 	if ( ! current || pos == s_len + 1 ) {
 	    // save result
-	    token_info.results[ field ] = s;
-	    token_info.lengths[ field ] = fieldlen - 1;
+	    token_info.results[ (int)field ] = s;
+	    token_info.lengths[ (int)field ] = fieldlen - 1;
 
 	    // end of string
 	    ret = 0;
@@ -177,8 +177,8 @@ static int qt_tokenize( const char *s, ulong s_len, ulong *advance,
 
 	if ( current == token_info.fields[ field ] ) {
 	    // save result
-	    token_info.results[ field ] = s;
-	    token_info.lengths[ field ] = fieldlen - 1;
+	    token_info.results[ (int)field ] = s;
+	    token_info.lengths[ (int)field ] = fieldlen - 1;
 
 	    // end of field
 	    fieldlen = 0;
